+++ src/test/java/com/aaronpmaus/jMath/transformations/TestTransformation.java
@@ -2,7 +2,7 @@ package com.aaronpmaus.jMath.transformations;
 
 import com.aaronpmaus.jMath.linearAlgebra.Matrix;
 import com.aaronpmaus.jMath.linearAlgebra.Vector;
-import com.aaronpmaus.jMath.linearAlgebra.Point3D;
+import com.aaronpmaus.jMath.linearAlgebra.Vector3D;
 
 import static org.junit.Assert.*;
 import org.junit.Test;
@@ -40,9 +40,9 @@ public class TestTransformation {
     Matrix expected = new Matrix(rot90);
     assertEquals(t.getMatrix(), expected);
 
-    Point3D coor = new Point3D(0.0, 1.0, 0.0);
+    Vector3D coor = new Vector3D(0.0, 1.0, 0.0);
     coor.applyTransformation(t);
-    expected = new Point3D(0.0,  0.0, 1.0);
+    expected = new Vector3D(0.0,  0.0, 1.0);
     assertEquals(coor, expected);
   }
 
@@ -57,9 +57,9 @@ public class TestTransformation {
     Matrix expected = new Matrix(rot90);
     assertEquals(t.getMatrix(), expected);
 
-    Point3D coor = new Point3D(0.0, 0.0, 1.0);
+    Vector3D coor = new Vector3D(0.0, 0.0, 1.0);
     coor.applyTransformation(t);
-    expected = new Point3D(1.0, 0.0, 0.0);
+    expected = new Vector3D(1.0, 0.0, 0.0);
     assertEquals(coor, expected);
   }
 
@@ -74,9 +74,9 @@ public class TestTransformation {
     Matrix expected = new Matrix(rot90);
     assertEquals(t.getMatrix(), expected);
 
-    Point3D coor = new Point3D(1.0, 0.0, 0.0);
+    Vector3D coor = new Vector3D(1.0, 0.0, 0.0);
     coor.applyTransformation(t);
-    expected = new Point3D(0.0, 1.0, 0.0);
+    expected = new Vector3D(0.0, 1.0, 0.0);
     assertEquals(coor, expected);
   }
 
@@ -85,7 +85,7 @@ public class TestTransformation {
     // Rotate about the X axis
     Transformation t = new Transformation();
     t.addRotationAboutAxis(
-        new Point3D(1.0, 0.0, 0.0), 90);
+        new Vector3D(1.0, 0.0, 0.0), 90);
     Double[][] rot90X = {{ 1.0,  0.0,  0.0,  0.0},
                         { 0.0,  0.0, -1.0,  0.0},
                         { 0.0,  1.0,  0.0,  0.0},
@@ -93,15 +93,15 @@ public class TestTransformation {
     Matrix expected = new Matrix(rot90X);
     assertEquals(t.getMatrix(), expected);
 
-    Point3D coor = new Point3D(0.0, 0.0, 1.0);
+    Vector3D coor = new Vector3D(0.0, 0.0, 1.0);
     coor.applyTransformation(t);
-    expected = new Point3D(0.0, -1.0, 0.0);
+    expected = new Vector3D(0.0, -1.0, 0.0);
     assertEquals(coor, expected);
 
     // Rotate about the Y axis
     t = new Transformation();
     t.addRotationAboutAxis(
-        new Point3D(0.0, 1.0, 0.0), 90);
+        new Vector3D(0.0, 1.0, 0.0), 90);
     Double[][] rot90Y = {{ 0.0,  0.0,  1.0,  0.0},
                         { 0.0,  1.0,  0.0,  0.0},
                         {-1.0,  0.0,  0.0,  0.0},
@@ -109,15 +109,15 @@ public class TestTransformation {
     expected = new Matrix(rot90Y);
     assertEquals(t.getMatrix(), expected);
 
-    coor = new Point3D(0.0, 0.0, 1.0);
+    coor = new Vector3D(0.0, 0.0, 1.0);
     coor.applyTransformation(t);
-    expected = new Point3D(1.0, 0.0, 0.0);
+    expected = new Vector3D(1.0, 0.0, 0.0);
     assertEquals(coor, expected);
 
     // Rotate about the Z axis
     t = new Transformation();
     t.addRotationAboutAxis(
-        new Point3D(0.0, 0.0, 1.0), 90);
+        new Vector3D(0.0, 0.0, 1.0), 90);
     Double[][] rot90Z = {{ 0.0, -1.0,  0.0,  0.0},
                         { 1.0,  0.0,  0.0,  0.0},
                         { 0.0,  0.0,  1.0,  0.0},
@@ -125,16 +125,16 @@ public class TestTransformation {
     expected = new Matrix(rot90Z);
     assertEquals(t.getMatrix(), expected);
 
-    coor = new Point3D(1.0, 0.0, 0.0);
+    coor = new Vector3D(1.0, 0.0, 0.0);
     coor.applyTransformation(t);
-    expected = new Point3D(0.0, 1.0, 0.0);
+    expected = new Vector3D(0.0, 1.0, 0.0);
     assertEquals(coor, expected);
   }
 
   @Test
   public void testTranslation(){
     Transformation t = new Transformation();
-    t.addTranslation(new Vector(3.0,4.0,5.0));
+    t.addTranslation(new Vector3D(3.0,4.0,5.0));
     Double[][] rot90 = {{ 1.0,  0.0,  0.0,  3.0},
                         { 0.0,  1.0,  0.0,  4.0},
                         { 0.0,  0.0,  1.0,  5.0},
@@ -142,9 +142,9 @@ public class TestTransformation {
     Matrix expected = new Matrix(rot90);
     assertEquals(t.getMatrix(), expected);
 
-    Point3D coor = new Point3D(1.0, 1.0, 1.0);
+    Vector3D coor = new Vector3D(1.0, 1.0, 1.0);
     coor.applyTransformation(t);
-    expected = new Point3D(4.0, 5.0, 6.0);
+    expected = new Vector3D(4.0, 5.0, 6.0);
     assertEquals(coor, expected);
   }
 
@@ -157,7 +157,7 @@ public class TestTransformation {
     t.addRotationAboutY(90);
     //System.out.println(t.getMatrix());
     //System.out.println("Adding Translation (6.0, 7.0, 8.0)");
-    t.addTranslation(new Vector(6.0,7.0,8.0));
+    t.addTranslation(new Vector3D(6.0,7.0,8.0));
     //System.out.println(t.getMatrix());
     Double[][] rot90 = {{ 0.0,  0.0,  1.0,  6.0},
                         { 0.0,  1.0,  0.0,  7.0},
@@ -167,11 +167,11 @@ public class TestTransformation {
     //System.out.println(expected);
     assertEquals(t.getMatrix(), expected);
 
-    Point3D coor = new Point3D(0.0, 0.0, 1.0);
+    Vector3D coor = new Vector3D(0.0, 0.0, 1.0);
     coor.applyTransformation(t);
     //System.out.println(coor);
     //System.out.println(expected);
-    expected = new Point3D(7.0,7.0,8.0);
+    expected = new Vector3D(7.0,7.0,8.0);
     assertEquals(coor, expected);
   }
 
@@ -180,16 +180,16 @@ public class TestTransformation {
     Transformation t = new Transformation();
     t.addRotationAboutY(90);
 
-    Point3D coor = new Point3D(0.0, 0.0, 1.0);
+    Vector3D coor = new Vector3D(0.0, 0.0, 1.0);
     coor.applyTransformation(t);
 
-    Matrix expected = new Point3D(1.0, 0.0, 0.0);
+    Matrix expected = new Vector3D(1.0, 0.0, 0.0);
     assertEquals(coor, expected);
 
     Transformation inverse = t.inverse();
 
     coor.applyTransformation(inverse);
-    expected = new Point3D(0.0, 0.0, 1.0);
+    expected = new Vector3D(0.0, 0.0, 1.0);
     assertEquals(coor, expected);
 
     t.addTransformation(inverse);
@@ -198,22 +198,22 @@ public class TestTransformation {
 
   @Test
   public void testComposedInverse(){
-    Point3D coor = new Point3D(0.0, 0.0, 1.0);
+    Vector3D coor = new Vector3D(0.0, 0.0, 1.0);
 
     Transformation t = new Transformation();
     t.addRotationAboutX(90); // After rotation: (0.0, -1.0, 0.0)
-    t.addTranslation(new Vector(43.0, 43.0, 43.0)); // After translation: (43.0, 42.0, 43.0)
+    t.addTranslation(new Vector3D(43.0, 43.0, 43.0)); // After translation: (43.0, 42.0, 43.0)
 
     // Apply the composed transformation and ensure the result is correct
     coor.applyTransformation(t);
-    Matrix expected = new Point3D(43.0, 42.0, 43.0);
+    Matrix expected = new Vector3D(43.0, 42.0, 43.0);
     assertEquals(coor, expected);
 
     Transformation inverse = t.inverse();
 
     // Apply the inverse and ensure the result is the original coordinates
     coor.applyTransformation(inverse);
-    expected = new Point3D(0.0, 0.0, 1.0);
+    expected = new Vector3D(0.0, 0.0, 1.0);
     assertEquals(coor, expected);
 
     // Apply the inverse to the original transformation and ensure the result is the identity
@@ -234,8 +234,8 @@ public class TestTransformation {
 
   @Test
   public void testRotateOntoVector(){
-    Point3D ref = new Point3D(1.0, 2.0, 3.0);
-    Point3D mobile = new Point3D(1.0, 1.0, 1.0);
+    Vector3D ref = new Vector3D(1.0, 2.0, 3.0);
+    Vector3D mobile = new Vector3D(1.0, 1.0, 1.0);
 
     Transformation t = new Transformation();
     t.addRotationOntoVector(ref, mobile);